% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runManual.R
\name{plotResults}
\alias{plotResults}
\title{#' Plot the results from the food web model or the trait-based model}
\usage{
plotResults(S)
}
\arguments{
\item{S}{Output from a previous simulation to use as initial conditions}

\item{Parameterset}{Text string with the name of the parameter set to use}

\item{F0}{Array with fishing mortality on each fleet}
}
\value{
-
}
\description{
#' Plot the results from the food web model or the trait-based model
}
\examples{
plotResults( runTraitBasedModel() )

}
