% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{runCommunitySpectrum}
\alias{runCommunitySpectrum}
\title{Run the community spectrum simulator}
\usage{
runCommunitySpectrum(...)
}
\arguments{
\item{...}{Not used at the moment}
}
\value{
Nothing, it is used to run the Shiny app
}
\description{
Run the community spectrum simulator
}
\examples{
runCommunitySpectrum()
}
