% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runManual.R
\name{runFoodWebModel}
\alias{runFoodWebModel}
\title{Run the food web size spectrum model manually}
\usage{
runFoodWebModel(Parameterset = "TraitBased", F0 = c(0.3, 0.3, 0.3),
  S = NA)
}
\arguments{
\item{Parameterset}{Text string with the name of the parameter set to use}

\item{F0}{Array with fishing mortality on each fleet}

\item{S}{Output from a previous simulation to use as initial conditions}
}
\value{
A structure with the output of the simulation
}
\description{
Run the food web size spectrum model manually
}
\examples{
runFoodWebModel('North Sea', F0 = c(0.1, 0.3, 0.7))


}
