% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runManual.R
\name{runTraitBasedModel}
\alias{runTraitBasedModel}
\title{Run the community spectrum simulator manually}
\usage{
runTraitBasedModel(nSpecies = 18, F0 = 0.3 + 0 * (1:nSpecies), S = NA,
  param = baseparameters(10^seq(log10(4), log10(30000), length.out =
  nSpecies), kappa = 1.3, h = 20))
}
\arguments{
\item{nSpecies}{Number of asymptotic size groups ("species").  Only relevant for trait-based model.}

\item{F0}{Array with fishing mortality on each species}

\item{S}{Output from a previous simulation to use as initial conditions}

\item{param}{Set of parameters to use}
}
\value{
A structure with the output of the simulation
}
\description{
Run the community spectrum simulator manually
}
\examples{
runTraitBasedModel()


}
